refactor(project-management): split the work-item atomic writer
